# resources/api_doc_template.yaml
swagger: "2.0"
info:
  description: "This API documentation details the resources for the Code Stroke Backend."
  version: "$VERSION"
  title: "Code Stroke Backend"
host: "$HOST"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /event_log/all/:
    get:
      tags:
      - "Event Log"
      summary: "View a list of all events."
      description: "Returns a list of all events in the database."
      operationId: "getEventsAll"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              success: 
                type: boolean
                default: True
              result:
                type: array
                items:
                  $ref: "#/definitions/EVENT_LOG"
        401:
          description: "Authentication failed."
          schema:
            $ref: "#/definitions/FailedAuthClinician"
      security:
        - requiresClinician: []
  /event_log/limit/:
    get:
      tags:
      - "Event Log"
      summary: "View a count-limited list of events."
      description: "Returns a list of a limited number of events in the database, from most recent to earliest."
      operationId: "getEventsLimit"
      produces:
      - "application/json"
      parameters:
        - in: "query"
          name: "start"
          type: "integer"
          required: True
          description: "The position of the first query e.g. if `start` is 25, then the queries will be returned starting from the 25th most recent, the 26th most recent etc."
        - in: "query"
          name: "number"
          type: "integer"
          required: True
          description: "The number of queries to return in total."
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            properties:
              success: 
                type: boolean
                default: True
              result:
                type: array
                items:
                  $ref: "#/definitions/EVENT_LOG"
        400:
          description: "Query parameters were invalid."
          schema:
            type: "object"
            properties:
              success:
                type: boolean
                default: False
              error_type:
                type: "string"
                enum: ["parameters"]
        401:
          description: "Authentication failed."
          schema:
            $ref: "#/definitions/FailedAuthClinician"
      security:
        - requiresClinician: []  
...
securityDefinitions:
  requiresClinician:
    type: "basic"
    description: "This is the standard 2FA method for a clinician. The time-based token based on the shared secret should be concatenated onto the password with a colon i.e. `password:token` - if using the headers directly, the whole string should be of the form `username:password:token` before being base64 encoded."
  requiresAdmin:
    type: "basic"
    description: "This route requires administrator credentials, which are just a standard username and password. These resources should not generally be called from the ordinary frontend clients and may be deprecated in the future."
definitions:
  FailedAuthClinician:
    type: "object"
    properties:
      success: 
        type: boolean
        default: False
      error_type:
        type: string
        enum: ["auth"]